Make modules actual `Modules`
We can build on top of the base `Purcel`.

We might provide an application specific purcel that has empty contents
for all of the modules.

Maybe there's some way to limit the boilerplate of defining an import?
A purcel could do that if so.

We could provide ready made purcels that you extend later.
Maybe they're a known set of modules that compile together like stackage.
Extending is trivial, you just use Dhall to add more dependencies.

We could provide templates to get people started easily.
These templates could be like a beginner level thing.
They could also be an internal company thing.
They could be a school project thing.

There are tons of possibilities.